+++ controllers/index.js
@@ -1,0 +1,8 @@
+
+const router = require('express').Router();
+
+const homeRoutes = require('./home-routes');
+
+router.use('/', homeRoutes);
+
+module.exports = router;
