+++ package.json
@@ -7,7 +7,7 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node server.js",
     "dev": "nodemon server.js",
-    "db:init": "mysql -u root  -p < db/schema.sql",
+    "db:init": "mysql -u kj -p < db/schema.sql",
     "seed": "node seeds/index.js"
   },
   "repository": {
